split functions into util
